+ scss module for ShBag and some refactor
