+++ src/components/Card/Card.js
@@ -28,47 +28,47 @@ const Card = ({
         onFavorite({title, price, imageUrl, id});
         setIsFavorite(!isFavorite);
     };
-    debugger
+
     return (
         <div className={styles.card}>
             {
                 isLoading ? <ContentLoader
-                    speed={2}
-                    width={150}
-                    height={250}
-                    viewBox="0 0 150 240"
-                    backgroundColor="#e0e0e0"
-                    foregroundColor="#ecebeb"
-                >
-                    <rect x="0" y="98" rx="0" ry="0" width="150" height="16"/>
-                    <rect x="4" y="59" rx="2" ry="2" width="142" height="26"/>
-                    <rect x="0" y="0" rx="10" ry="10" width="150" height="90"/>
-                    <rect x="-6" y="124" rx="0" ry="0" width="100" height="23"/>
-                    <rect x="1" y="205" rx="0" ry="0" width="80" height="24"/>
-                    <rect x="113" y="200" rx="8" ry="8" width="32" height="32"/>
-                </ContentLoader>
-                :
-                <>
-                    <div className={styles.favorite}>
-                        <img
-                            onClick={onClickFavorites}
-                            src={isFavorite ? heartLiked : heartUnliked} alt="heartUnliked" width={20}/>
-                    </div>
-                    <img src={imageUrl} alt="sneakers" width={133} height={112}/>
-                    <h5>{title}</h5>
-                    <div className='d-flex justify-between align-center'>
-                        <div className='d-flex flex-column '>
-                            <span>Price: </span>
-                            <b>{price}$</b>
+                        speed={2}
+                        width={150}
+                        height={250}
+                        viewBox="0 0 150 240"
+                        backgroundColor="#e0e0e0"
+                        foregroundColor="#ecebeb"
+                    >
+                        <rect x="0" y="98" rx="0" ry="0" width="150" height="16"/>
+                        <rect x="4" y="59" rx="2" ry="2" width="142" height="26"/>
+                        <rect x="0" y="0" rx="10" ry="10" width="150" height="90"/>
+                        <rect x="-6" y="124" rx="0" ry="0" width="100" height="23"/>
+                        <rect x="1" y="205" rx="0" ry="0" width="80" height="24"/>
+                        <rect x="113" y="200" rx="8" ry="8" width="32" height="32"/>
+                    </ContentLoader>
+                    :
+                    <>
+                        {onFavorite && <div className={styles.favorite}>
+                            <img
+                                onClick={onClickFavorites}
+                                src={isFavorite ? heartLiked : heartUnliked} alt="heartUnliked" width={20}/>
+                        </div>}
+                        <img src={imageUrl} alt="sneakers" width={133} height={112}/>
+                        <h5>{title}</h5>
+                        <div className='d-flex justify-between align-center'>
+                            <div className='d-flex flex-column '>
+                                <span>Price: </span>
+                                <b>{price}$</b>
+                            </div>
+
+                            {onPlus && <img src={isItemAdded(id) ? checked : plus} alt="plus"
+                                            className={styles.button}
+
+                                            onClick={onClickPlus}
+                            />}
                         </div>
-
-                        <img src={isItemAdded(id) ? checked : plus } alt="plus"
-                             className={styles.button}
-
-                             onClick={onClickPlus}
-                        />
-                    </div>
-                </>
+                    </>
             }
