+++ src/components/shoppingBag/ShoppingBag.module.scss
@@ -1,0 +1,91 @@
+
+.overlay {
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.71);
+  z-index: 1;
+  visibility: hidden;
+  opacity: 0;
+  transition: opacity 0.3s ease-in-out, visibility 0.3s ease-in-out;
+  overflow: hidden;
+}
+
+.overlayVisible {
+  opacity: 1;
+  visibility: visible;
+
+  .drawer {
+    transform: translateX(0);
+  }
+}
+
+.drawer {
+  position: absolute;
+  width: 420px;
+  min-height: 100%;
+  right: 0;
+  background: white;
+  box-shadow: -10px, 4px 24px rgba(0, 0, 0, 0.1);
+  padding: 30px;
+  display: flex;
+  flex-direction: column;
+  transform: translateX(100%);
+  transition: transform 0.3s ease-in-out;
+
+  h2 {
+    margin: 0;
+  }
+
+  .items {
+    flex: 1;
+    overflow: auto;
+    margin-bottom: 20px;
+    align-self: flex-end;
+  }
+
+
+}
+
+
+
+  .greenBtn {
+    position: relative;
+    width: 100%;
+    height: 55px;
+    background: #9dd558;
+    border-radius: 18px;
+    border: 0;
+    color: white;
+    font-size: 16px;
+    font-weight: 600;
+    cursor: pointer;
+    transition: background 0.2s ease-in-out;
+
+    &:disabled {
+      background-color: #bebebe;
+      cursor: default;
+    }
+
+    &:hover {
+      background: lighten(#9dd558, 10%);
+
+      .arrowLeft {
+        transform: translateX(-10px);
+      }
+
+    }
+
+    &:active {
+      background: darken(#9dd558, 10%);
+    }
+
+    .arrowLeft {
+      position: absolute;
+      left: 30px;
+      transition: transform 0.2s ease-in-out;
+
+    }
+  }
