+++ src/index.scss
@@ -78,104 +78,33 @@ header {
   }
 }
 
-.overlay {
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.71);
-  z-index: 1;
-}
 
 
-.drawer {
-  position: absolute;
-  width: 420px;
-  min-height: 100%;
-  right: 0;
-  background: white;
-  box-shadow: -10px, 4px 24px rgba(0, 0, 0, 0.1);
-  padding: 30px;
-  display: flex;
-  flex-direction: column;
+.downBlock {
 
-  h2 {
-    margin: 0;
-  }
-
-  .items {
-    flex: 1;
-    overflow: auto;
-    margin-bottom: 20px;
-    align-self: flex-end;
-  }
-
-  .downBlock {
-
-    .cartTotalBlock {
-      ul {
-        display: flex;
-        margin-bottom: 40px !important;
-      }
-
-      li {
-        display: flex;
-        align-items: flex-end;
-        margin-bottom: 20px;
-
-        div {
-          flex: 1;
-          height: 1px;
-          border-bottom: 1px dashed grey;
-          position: relative;
-          top: -4px;
-          margin: 0 10px;
-        }
-      }
-
+  .cartTotalBlock {
+    ul {
+      display: flex;
+      margin-bottom: 40px !important;
     }
 
-    .greenBtn {
-      position: relative;
-      width: 100%;
-      height: 55px;
-      background: #9dd558;
-      border-radius: 18px;
-      border: 0;
-      color: white;
-      font-size: 16px;
-      font-weight: 600;
-      cursor: pointer;
-      transition: background 0.2s ease-in-out;
-      &:disabled {
-        background-color: #bebebe;
-        cursor: default;
-      }
-      &:hover {
-        background: lighten(#9dd558, 10%);
+    li {
+      display: flex;
+      align-items: flex-end;
+      margin-bottom: 20px;
 
-        .arrow {
-          transform: translateX(10px);
-        }
-
-      }
-
-      &:active {
-        background: darken(#9dd558, 10%);
-      }
-
-      .arrow {
-        position: absolute;
-        right: 30px;
-        transition: transform 0.2s ease-in-out;
-
+      div {
+        flex: 1;
+        height: 1px;
+        border-bottom: 1px dashed grey;
+        position: relative;
+        top: -4px;
+        margin: 0 10px;
       }
     }
-  }
-}
 
-.cartEmpty {
+  }
+
   .greenBtn {
     position: relative;
     width: 100%;
@@ -197,8 +126,8 @@ header {
     &:hover {
       background: lighten(#9dd558, 10%);
 
-      .arrowLeft {
-        transform: translateX(-10px);
+      .arrow {
+        transform: translateX(10px);
       }
 
     }
@@ -207,14 +136,54 @@ header {
       background: darken(#9dd558, 10%);
     }
 
-    .arrowLeft {
+    .arrow {
       position: absolute;
-      left: 30px;
+      right: 30px;
       transition: transform 0.2s ease-in-out;
 
     }
   }
 }
+
+.greenBtn {
+  position: relative;
+  width: 100%;
+  height: 55px;
+  background: #9dd558;
+  border-radius: 18px;
+  border: 0;
+  color: white;
+  font-size: 16px;
+  font-weight: 600;
+  cursor: pointer;
+  transition: background 0.2s ease-in-out;
+
+  &:disabled {
+    background-color: #bebebe;
+    cursor: default;
+  }
+
+  &:hover {
+    background: lighten(#9dd558, 10%);
+
+    .arrowLeft {
+      transform: translateX(-10px);
+    }
+
+  }
+
+  &:active {
+    background: darken(#9dd558, 10%);
+  }
+
+  .arrowLeft {
+    position: absolute;
+    left: 30px;
+    transition: transform 0.2s ease-in-out;
+
+  }
+}
+
 
 
 .cartItem {
